requirements.txt
streamlit
google.generativeai
pypdf
python-docx
Pillow
opencv-python